top2mid: support Modulation commands
I reworked the way the RunningNotes queue system works, in order to make
"sound cut" notes less hackish.
I also planned to add Modulation CCs into the queue, but in the end
Modulation support was done similarly to pitch slide support in mdc2mid,
because that made it easier to keep track of the current state.